Add helper method instead of repeating code.
