+++ includes/core/classes/class-venue-setup.php
@@ -17,6 +17,7 @@ namespace GatherPress\Core;
 defined( 'ABSPATH' ) || exit; // @codeCoverageIgnore
 
 use GatherPress\Core\Traits\Singleton;
+use GatherPress\Core\Utility;
 use GatherPress\Core\Validate;
 use stdClass;
 use WP_Block_Patterns_Registry;
@@ -221,35 +222,6 @@ class Venue_Setup {
 	}
 
 	/**
-	 * Authorization callback for post meta that mirrors the post-level edit cap.
-	 *
-	 * Routes through `user_can( $user_id, 'edit_post', $object_id )` so the
-	 * per-post permission model (`map_meta_cap` → `edit_others_posts`,
-	 * `edit_published_posts`, etc.) gates meta the same way it gates the post
-	 * itself. Without this, the meta layer would be more permissive than the
-	 * post layer that owns it: a custom REST route or third-party
-	 * `update_post_meta()` call could bypass the per-post check that the WP
-	 * posts controller already enforces on the post.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param bool   $allowed   Whether the user can edit the post meta. Unused;
-	 *                          we authoritatively return based on `edit_post`.
-	 * @param string $meta_key  The meta key being accessed. Unused.
-	 * @param int    $object_id The post ID the meta belongs to.
-	 * @param int    $user_id   The user ID attempting the edit.
-	 * @return bool True if the user can edit the post, false otherwise.
-	 */
-	public function can_edit_post_meta(
-		bool $allowed,
-		string $meta_key,
-		int $object_id,
-		int $user_id
-	): bool {
-		return user_can( $user_id, 'edit_post', $object_id );
-	}
-
-	/**
 	 * Sanitize callback for venue coordinate meta (latitude / longitude).
 	 *
 	 * Numeric values within the ±180 range are normalized to a string form via
@@ -283,7 +255,7 @@ class Venue_Setup {
 	public function maybe_register_post_meta( string $post_type ): void {
 		$venue_information_meta = array(
 			'gatherpress_address'   => array(
-				'auth_callback'     => array( $this, 'can_edit_post_meta' ),
+				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
 				'sanitize_callback' => 'sanitize_text_field',
 				'show_in_rest'      => true,
 				'single'            => true,
@@ -292,7 +264,7 @@ class Venue_Setup {
 				'revisions_enabled' => true,
 			),
 			'gatherpress_latitude'  => array(
-				'auth_callback'     => array( $this, 'can_edit_post_meta' ),
+				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
 				'sanitize_callback' => array( $this, 'sanitize_coordinate' ),
 				'show_in_rest'      => true,
 				'single'            => true,
@@ -301,7 +273,7 @@ class Venue_Setup {
 				'revisions_enabled' => true,
 			),
 			'gatherpress_longitude' => array(
-				'auth_callback'     => array( $this, 'can_edit_post_meta' ),
+				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
 				'sanitize_callback' => array( $this, 'sanitize_coordinate' ),
 				'show_in_rest'      => true,
 				'single'            => true,
@@ -310,7 +282,7 @@ class Venue_Setup {
 				'revisions_enabled' => true,
 			),
 			'gatherpress_phone'     => array(
-				'auth_callback'     => array( $this, 'can_edit_post_meta' ),
+				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
 				'sanitize_callback' => 'sanitize_text_field',
 				'show_in_rest'      => true,
 				'single'            => true,
@@ -319,7 +291,7 @@ class Venue_Setup {
 				'revisions_enabled' => true,
 			),
 			'gatherpress_website'   => array(
-				'auth_callback'     => array( $this, 'can_edit_post_meta' ),
+				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
 				'sanitize_callback' => 'sanitize_url',
 				'show_in_rest'      => true,
 				'single'            => true,
@@ -354,7 +326,7 @@ class Venue_Setup {
 		$venue_map_meta = array(
 			// Map display settings.
 			'gatherpress_venue_map_show'   => array(
-				'auth_callback'     => array( $this, 'can_edit_post_meta' ),
+				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
 				'sanitize_callback' => 'rest_sanitize_boolean',
 				'show_in_rest'      => true,
 				'single'            => true,
@@ -362,7 +334,7 @@ class Venue_Setup {
 				'default'           => true,
 			),
 			'gatherpress_venue_map_zoom'   => array(
-				'auth_callback'     => array( $this, 'can_edit_post_meta' ),
+				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
 				'sanitize_callback' => 'absint',
 				'show_in_rest'      => true,
 				'single'            => true,
@@ -370,7 +342,7 @@ class Venue_Setup {
 				'default'           => 10,
 			),
 			'gatherpress_venue_map_height' => array(
-				'auth_callback'     => array( $this, 'can_edit_post_meta' ),
+				'auth_callback'     => array( Utility::class, 'can_edit_post_meta' ),
 				'sanitize_callback' => 'absint',
 				'show_in_rest'      => true,
 				'single'            => true,
