+++ test/unit/php/includes/tests/core/classes/class-test-venue-setup.php
@@ -339,56 +339,6 @@ class Test_Venue_Setup extends Base {
 	}
 
 	/**
-	 * Tests can_edit_post_meta authorization callback — per-post check.
-	 *
-	 * Routes through `user_can( $user_id, 'edit_post', $object_id )` so the
-	 * permission model matches what WP uses for the post itself: Editors
-	 * (with `edit_others_posts`) can edit anyone's venue meta, Authors can
-	 * edit only their own, Subscribers and logged-out users are denied.
-	 *
-	 * @covers ::can_edit_post_meta
-	 *
-	 * @return void
-	 */
-	public function test_can_edit_post_meta(): void {
-		$instance = Venue_Setup::get_instance();
-
-		$author_one_id = $this->factory->user->create( array( 'role' => 'author' ) );
-		$author_two_id = $this->factory->user->create( array( 'role' => 'author' ) );
-		$editor_id     = $this->factory->user->create( array( 'role' => 'editor' ) );
-		$subscriber_id = $this->factory->user->create( array( 'role' => 'subscriber' ) );
-
-		$venue_id = $this->factory->post->create(
-			array(
-				'post_type'   => Venue::POST_TYPE,
-				'post_author' => $author_one_id,
-				'post_status' => 'publish',
-			)
-		);
-
-		$this->assertTrue(
-			$instance->can_edit_post_meta( false, 'gatherpress_address', $venue_id, $author_one_id ),
-			'The venue author should be able to edit their own venue meta.'
-		);
-		$this->assertFalse(
-			$instance->can_edit_post_meta( false, 'gatherpress_address', $venue_id, $author_two_id ),
-			'A different author should not be able to edit a venue they do not own.'
-		);
-		$this->assertTrue(
-			$instance->can_edit_post_meta( false, 'gatherpress_address', $venue_id, $editor_id ),
-			'An editor should be able to edit any venue meta via edit_others_posts.'
-		);
-		$this->assertFalse(
-			$instance->can_edit_post_meta( false, 'gatherpress_address', $venue_id, $subscriber_id ),
-			'A subscriber should not be able to edit any venue meta.'
-		);
-		$this->assertFalse(
-			$instance->can_edit_post_meta( false, 'gatherpress_address', $venue_id, 0 ),
-			'A logged-out user (user_id 0) should not be able to edit any venue meta.'
-		);
-	}
-
-	/**
 	 * Coverage for sanitize_coordinate.
 	 *
 	 * Numeric values within the ±180 range pass through; everything else
