+++ test/unit/php/includes/tests/core/classes/event/class-test-setup.php
@@ -1112,7 +1112,6 @@ class Test_Setup extends Base {
 	 *
 	 * @covers ::register_event_only_meta
 	 * @covers ::maybe_register_event_date_meta
-	 * @covers ::can_edit_post_meta
 	 * @return void
 	 */
 	public function test_register_post_meta_auth_callbacks(): void {
@@ -1149,56 +1148,6 @@ class Test_Setup extends Base {
 
 		$result = update_post_meta( $post_id, 'gatherpress_max_attendance_limit', 100 );
 		$this->assertNotFalse( $result, 'Should allow meta update for max_attendance_limit' );
-	}
-
-	/**
-	 * Tests can_edit_post_meta authorization callback — per-post check.
-	 *
-	 * Routes through `user_can( $user_id, 'edit_post', $object_id )` so the
-	 * permission model matches what WP uses for the post itself: Editors
-	 * (with `edit_others_posts`) can edit anyone's event meta, Authors can
-	 * edit only their own, Subscribers and logged-out users are denied.
-	 *
-	 * @covers ::can_edit_post_meta
-	 *
-	 * @return void
-	 */
-	public function test_can_edit_post_meta(): void {
-		$instance = Setup::get_instance();
-
-		$author_one_id = $this->factory->user->create( array( 'role' => 'author' ) );
-		$author_two_id = $this->factory->user->create( array( 'role' => 'author' ) );
-		$editor_id     = $this->factory->user->create( array( 'role' => 'editor' ) );
-		$subscriber_id = $this->factory->user->create( array( 'role' => 'subscriber' ) );
-
-		$event_id = $this->factory->post->create(
-			array(
-				'post_type'   => Event::POST_TYPE,
-				'post_author' => $author_one_id,
-				'post_status' => 'publish',
-			)
-		);
-
-		$this->assertTrue(
-			$instance->can_edit_post_meta( false, 'gatherpress_datetime', $event_id, $author_one_id ),
-			'The event author should be able to edit their own event meta.'
-		);
-		$this->assertFalse(
-			$instance->can_edit_post_meta( false, 'gatherpress_datetime', $event_id, $author_two_id ),
-			'A different author should not be able to edit an event they do not own.'
-		);
-		$this->assertTrue(
-			$instance->can_edit_post_meta( false, 'gatherpress_datetime', $event_id, $editor_id ),
-			'An editor should be able to edit any event meta via edit_others_posts.'
-		);
-		$this->assertFalse(
-			$instance->can_edit_post_meta( false, 'gatherpress_datetime', $event_id, $subscriber_id ),
-			'A subscriber should not be able to edit any event meta.'
-		);
-		$this->assertFalse(
-			$instance->can_edit_post_meta( false, 'gatherpress_datetime', $event_id, 0 ),
-			'A logged-out user (user_id 0) should not be able to edit any event meta.'
-		);
 	}
 
 	/**
